fix:Fix the way items in the  legend of the doughnut chart are display
Fix the number items that should show to get better aproach of the space
At the moment the items display but some cases don't got the correct space
